Refactor engine into modules
