+++ src/constants.rs
@@ -1,0 +1,2 @@
+pub const KING_VALUE: i32 = 20000;
+pub const MATE_SCORE: i32 = KING_VALUE * 10;
