+++ src/uci.rs
@@ -1,0 +1,125 @@
+use std::io::{self, BufRead, Write};
+use std::time::{Duration, Instant};
+
+use crate::board::{find_best_move_with_time, parse_uci_move, Board};
+use crate::transposition_table::TranspositionTable;
+use crate::types::{format_square, Move, Piece};
+
+pub fn parse_position_command(board: &mut Board, parts: &[&str]) {
+    let mut i = 1;
+    if i < parts.len() && parts[i] == "startpos" {
+        *board = Board::new();
+        i += 1;
+    } else if i < parts.len() && parts[i] == "fen" {
+        let fen = parts[i + 1..i + 7].join(" ");
+        *board = Board::from_fen(&fen);
+        i += 7;
+    }
+
+    if i < parts.len() && parts[i] == "moves" {
+        i += 1;
+        while i < parts.len() {
+            if let Some(mv) = parse_uci_move(board, parts[i]) {
+                board.make_move(&mv);
+            } else {
+                eprintln!("Invalid move: {}", parts[i]);
+            }
+            i += 1;
+        }
+    }
+}
+
+pub fn format_uci_move(mv: &Move) -> String {
+    let mut s = format!("{}{}", format_square(mv.from), format_square(mv.to));
+    if let Some(promo) = mv.promotion {
+        s.push(match promo {
+            Piece::Queen => 'q',
+            Piece::Rook => 'r',
+            Piece::Bishop => 'b',
+            Piece::Knight => 'n',
+            _ => '?',
+        });
+    }
+    s
+}
+
+pub fn run_uci_loop() {
+    let stdin = io::stdin();
+    let mut stdout = io::stdout();
+    let mut board = Board::new();
+    let mut tt = TranspositionTable::new(1024);
+
+    let mut time_left = Duration::from_secs(5);
+    let mut inc = Duration::ZERO;
+    let mut movetime: Option<Duration> = None;
+
+    for line in stdin.lock().lines() {
+        let line = line.unwrap();
+        let parts: Vec<&str> = line.trim().split_whitespace().collect();
+        if parts.is_empty() {
+            continue;
+        }
+
+        match parts[0] {
+            "uci" => {
+                println!("id name MyRustEngine");
+                println!("id author Dean Menezes");
+                println!("uciok");
+            }
+            "isready" => {
+                println!("readyok");
+            }
+            "ucinewgame" => {
+                board = Board::new();
+            }
+            "position" => {
+                parse_position_command(&mut board, &parts);
+            }
+            "go" => {
+                let mut i = 1;
+                while i < parts.len() {
+                    match parts[i] {
+                        "wtime" if board.white_to_move => {
+                            time_left = Duration::from_millis(parts[i + 1].parse().unwrap_or(1000));
+                            i += 2;
+                        }
+                        "btime" if !board.white_to_move => {
+                            time_left = Duration::from_millis(parts[i + 1].parse().unwrap_or(1000));
+                            i += 2;
+                        }
+                        "winc" if board.white_to_move => {
+                            inc = Duration::from_millis(parts[i + 1].parse().unwrap_or(0));
+                            i += 2;
+                        }
+                        "binc" if !board.white_to_move => {
+                            inc = Duration::from_millis(parts[i + 1].parse().unwrap_or(0));
+                            i += 2;
+                        }
+                        "movetime" => {
+                            movetime =
+                                Some(Duration::from_millis(parts[i + 1].parse().unwrap_or(100)));
+                            i += 2;
+                        }
+                        _ => i += 1,
+                    }
+                }
+
+                let max_time = movetime.unwrap_or_else(|| time_left / 30 + inc);
+                let start = Instant::now();
+
+                if let Some(best_move) =
+                    find_best_move_with_time(&mut board, &mut tt, max_time, start)
+                {
+                    let uci_move = format_uci_move(&best_move);
+                    println!("bestmove {}", uci_move);
+                } else {
+                    println!("bestmove 0000");
+                }
+            }
+            "quit" => break,
+            _ => {}
+        }
+
+        stdout.flush().unwrap();
+    }
+}
